requirements.txt
fastapi<0.100
python-dotenv
python-multipart
Pillow
fastapi-jwt-auth==0.5.0
pydantic<2
alembic
uvicorn
sqlalchemy
sqlalchemy-utils
psycopg2-binary
bcrypt
